+++ packages/FormElement/src/components/Fieldset/index.js
@@ -1,7 +1,1 @@
-export { default as Content } from "../Content";
-export { default as Description } from "../Description";
-export { default as Error } from "../Error";
-export { default as Help } from "../Help";
-export { default as Instructions } from "../Instructions";
-export { default as Label } from "../Label";
 export { default } from "./Fieldset";