+++ packages/FormElement/src/components/Layout/LeftCol/LeftCol.js
@@ -3,22 +3,16 @@ import PropTypes from "prop-types";
 import * as sc from "./LeftCol.styles";
 
 const propTypes = {
-  width: PropTypes.string,
   children: PropTypes.node,
 };
 
 const defaultProps = {
-  width: "180px",
   children: null,
 };
 
 function LeftCol(props) {
-  const { children, width, ...moreProps } = props;
-  return (
-    <sc.LeftCol width={width} {...moreProps}>
-      {children}
-    </sc.LeftCol>
-  );
+  const { children, ...moreProps } = props;
+  return <sc.LeftCol {...moreProps}>{children}</sc.LeftCol>;
 }
 
 LeftCol.displayName = "FormElement.Layout.LeftCol";