+++ packages/FormElement/src/components/Layout/LeftCol/LeftCol.styles.js
@@ -2,6 +2,7 @@ import styled, { css } from "styled-components";
 
 export const LeftCol = styled.div(
   ({ width }) => css`
+    flex-shrink: 0;
     width: ${width};
   `
 );