+++ packages/FormElement/stories/examples/Checkbox.js
@@ -3,9 +3,9 @@ import * as constants from "@paprika/constants/lib/Constants";
 import Checkbox from "@paprika/checkbox";
 import styled from "styled-components";
 import stylers from "@paprika/stylers";
-import FormElement from "../../src";
+import { Fieldset } from "../../src";
 
-const { Label, Description, Content } = FormElement;
+const { Label, Description, Content } = Fieldset;
 
 export default function CheckboxExample() {
   const isDisabled = false;
@@ -25,10 +25,10 @@ export default function CheckboxExample() {
   `;
 
   return (
-    <FormElement hasFieldSet>
+    <Fieldset>
       <Label>Form Label</Label>
       <Description>A description</Description>
       <Content>{a11yProps => <CheckboxGroup>{getCheckboxOptions(a11yProps)}</CheckboxGroup>}</Content>
-    </FormElement>
+    </Fieldset>
   );
 }