+++ packages/FormElement/stories/examples/Radio.js
@@ -3,9 +3,9 @@ import * as constants from "@paprika/constants/lib/Constants";
 import { action } from "@storybook/addon-actions";
 import Radio from "@paprika/radio";
 import stylers from "@paprika/stylers";
-import FormElement from "../../src";
+import { Fieldset } from "../../src";
 
-const { Content, Label } = FormElement;
+const { Content, Label, Description } = Fieldset;
 
 export default function RadioExample() {
   const optionsArray = ["Black Panther", "Wonder Woman", "Spiderman", "The Incredibles", "Thor", <span>test</span>];
@@ -20,16 +20,18 @@ export default function RadioExample() {
       }}
     >
       {optionsArray.map(hero => (
-        <Radio key={hero} isDisabled={isDisabled} size={size} {...a11yProps}>
+        <Radio key={hero} isDisabled={isDisabled} size={size}>
+          <Radio.Input {...a11yProps} />
           {hero}
         </Radio>
       ))}
     </Radio.Group>
   );
   return (
-    <FormElement hasFieldSet>
+    <Fieldset>
       <Label>Form Label</Label>
+      <Description>A description</Description>
       <Content>{a11yProps => renderRadioGroup(a11yProps)}</Content>
-    </FormElement>
+    </Fieldset>
   );
 }