+++ packages/Radio/src/Radio.js
@@ -2,14 +2,15 @@ import React from "react";
 import PropTypes from "prop-types";
 import nanoid from "nanoid";
 import CheckIcon from "@paprika/icon/lib/Check";
+import extractChildrenProps from "@paprika/helpers/lib/extractChildrenProps";
 import * as sc from "./Radio.styles";
 import Group from "./components/Group";
+import RadioInputPropsCollector from "./RadioInputPropsCollector";
 import types from "./types";
 
 function Radio(props) {
   const {
     a11yText,
-    ariaDescribedBy,
     children,
     isChecked,
     isDisabled,
@@ -23,6 +24,11 @@ function Radio(props) {
   } = props;
   const radioId = React.useRef(nanoid()).current;
   const inputRef = React.useRef(null);
+  const extendedInputProps = extractChildrenProps(children, RadioInputPropsCollector);
+
+  if (extendedInputProps.refLabel) {
+    delete extendedInputProps.refLabel;
+  }
 
   const handleKeyDown = event => {
     if (
@@ -48,7 +54,7 @@ function Radio(props) {
   };
 
   const inputProps = {
-    "aria-describedby": ariaDescribedBy,
+    ...extendedInputProps,
     onClick,
     checked: isChecked,
     disabled: isDisabled,
@@ -86,8 +92,6 @@ Radio.types = types;
 const propTypes = {
   /** Used for aria-label on the radio input  */
   a11yText: PropTypes.string,
-  /** Used for aria-describedby on the radio input  */
-  ariaDescribedBy: PropTypes.string,
   /** Describe if the radio started as selected or not */
   canDeselect: PropTypes.bool,
   /** Used for label contents */
@@ -112,7 +116,6 @@ const propTypes = {
 
 const defaultProps = {
   a11yText: null,
-  ariaDescribedBy: null,
   canDeselect: false,
   children: null,
   defaultIsChecked: false,
@@ -129,5 +132,7 @@ Radio.displayName = "Radio";
 Radio.propTypes = propTypes;
 Radio.defaultProps = defaultProps;
 Radio.Group = Group;
+
+Radio.Input = RadioInputPropsCollector;
 
 export default Radio;