+++ packages/FormElement/CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
-- Refactor FormElement api [@tristanjasper](https://github.com/tristanjasper).
+- Refactor FormElement api to use a11yProps instead of seperate attributes, Additionally using context over React cloning. Introduced Fieldset and Layout components [@tristanjasper](https://github.com/tristanjasper).
 
 ## [0.3.0] - 2020-01-17